refactor(sdk): extract ENS recipient utils and move resolved swap types
